Better debug logs on response (#19595)
* Output information about what's being broadcast

* Expose/promote QNameString in Logging to allow other logging locations to get a formatted FullQName

* Revert inadvertent build_overrides/pigweed_environment.gni

This partially reverts commit 15639cf3c5ef3c21480e68b11bdbcd39012a7a11.

# Conflicts:
#	build_overrides/pigweed_environment.gni

* Update BUILD.gn deps for QNameString & move it to its own file to avoid circular deps in /responders

* Remove unnecessary StringBuilder addition as QNameString is now in its own file

* Restyled by whitespace

* Restyled by clang-format
